requirements.txt
flask>=3.1.0
ebooklib>=0.18
PyPDF2>=3.0.1
openai>=1.60.2
python-dotenv>=1.0.1